requirements.txt
asgiref==3.7.1
Django==4.2.1
gunicorn==21.2.0
# gevent==22.10.2
pytz==2022.7.1
django-taggit==4.0.0
Markdown==3.4.3
Pillow==9.5.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
tzdata==2023.3
django-widget-tweaks
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
sorl-thumbnail==12.7.0
django-filter==23.2
requests==2.26.0